improve open in app link for existing LBRY users
